Remove the `Vec<Job>` from `DependencyQueue`
I... don't think this needed any more! Doing some digging looks like
this was originally added in 91a15e724. That was so early it didn't even
use a PR and was almost 5 years ago. Since then we've had a huge number
of changes to the backend and `Unit` nowadays does all the deduplication
we need, so no need to store a `Vec` here and we can just have a mapping
of `Key` to `Job` and that's it.