Only execute `exclude_from_backups` once
This was currently getting executed on all builds, even if the directory
already exists. There shouldn't be any reason though to exclude the
directory from backups on all builds, and after seeing this get a stack
sample in a profile I figured it's best to ensure it only executes once
in case the backing system implementation isn't the speediest.