add warning to cargo incase clippy is missing
This is only a warning because rustup(?) already warns and is much more
verbose about how to recover from the error, so I would like to call
into the clippy-driver shim or w/e is outputting nice error messages.

But incase rustup is not available or some other shenanigans occur,
cargo will also output the warning so the error isn't a mystery.